# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/search/{city}": {
            "get": {
                "description": "Get list of cities by given City name.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "City Name",
                        "name": "city",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/weather/{city}/{country}/{lat}/{lon}": {
            "get": {
                "description": "Get data of location by given Lat \u0026 Lon.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "City",
                        "name": "city",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Country",
                        "name": "country",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Latitude",
                        "name": "lat",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Longitude",
                        "name": "lon",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    }
}